requirements.txt
flask
sqlalchemy
python-dotenv==1.0.1
